requirements.txt
# Backend
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.35.0
pandas>=2.1.0
plotly>=5.18.0
altair>=5.2.0
openpyxl>=3.1.0
matplotlib>=3.7.0
extra-streamlit-components>=0.1.60
streamlit-echarts>=0.4.0
streamlit-flow-component>=1.6.0
pyvis>=0.3.2

# Shared
requests>=2.31.0
python-dateutil>=2.8.2

# Cache
diskcache>=5.6.3

# PDF generation
reportlab>=4.0
fpdf>=1.7.2
PyPDF2>=3.0.0

# BUENOBOT - QA & Security Tools
filelock>=3.12.0
ruff>=0.1.0
pip-audit>=2.6.0
bandit>=1.7.0
jsonpath-ng>=1.6.0
PyYAML>=6.0
